+++ src/test/java/com/google/jenkins/plugins/computeengine/ComputeEngineCloudWindowsIT.java
@@ -1,7 +1,10 @@
 package com.google.jenkins.plugins.computeengine;
 
 import com.cloudbees.jenkins.plugins.sshcredentials.impl.BasicSSHUserPrivateKey;
-import com.cloudbees.plugins.credentials.*;
+import com.cloudbees.plugins.credentials.Credentials;
+import com.cloudbees.plugins.credentials.CredentialsScope;
+import com.cloudbees.plugins.credentials.CredentialsStore;
+import com.cloudbees.plugins.credentials.SystemCredentialsProvider;
 import com.cloudbees.plugins.credentials.common.StandardUsernameCredentials;
 import com.cloudbees.plugins.credentials.domains.Domain;
 import com.cloudbees.plugins.credentials.domains.DomainRequirement;
@@ -20,15 +23,24 @@ import hudson.model.Node;
 import hudson.model.Result;
 import hudson.model.labels.LabelAtom;
 import hudson.slaves.NodeProvisioner;
+import hudson.tasks.BatchFile;
 import hudson.tasks.Builder;
-import hudson.tasks.Shell;
+
 import org.awaitility.Awaitility;
-import org.junit.*;
+import org.junit.After;
+import org.junit.AfterClass;
+import org.junit.BeforeClass;
+import org.junit.ClassRule;
+import org.junit.Test;
 import org.jvnet.hudson.test.JenkinsRule;
 
 import java.io.ByteArrayOutputStream;
 import java.io.IOException;
-import java.util.*;
+import java.util.ArrayList;
+import java.util.Collection;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
 import java.util.concurrent.TimeUnit;
 import java.util.logging.LogManager;
 import java.util.logging.Logger;
@@ -63,8 +75,6 @@ public class ComputeEngineCloudWindowsIT {
     private static final String ACCELERATOR_NAME = "";
     private static final String ACCELERATOR_COUNT = "";
     private static final String RUN_AS_USER = "jenkins";
-
-    private static final String WINDOWS_USER = "Build";
 
     private static Map<String, String> INTEGRATION_LABEL;
 
@@ -135,7 +145,7 @@ public class ComputeEngineCloudWindowsIT {
         // Have to reformat since GoogleKeyPair's format is for metadata server and not typical public key format
         publicKey = kp.getPublicKey().trim().substring(1);
 
-        StandardUsernameCredentials windowsPrivateKeyCredential = new BasicSSHUserPrivateKey(CredentialsScope.GLOBAL, null, WINDOWS_USER,
+        StandardUsernameCredentials windowsPrivateKeyCredential = new BasicSSHUserPrivateKey(CredentialsScope.GLOBAL, null, RUN_AS_USER,
                 new BasicSSHUserPrivateKey.DirectEntryPrivateKeySource(kp.getPrivateKey()), null,
                 "integration test private key for windows");
         store.addCredentials(Domain.global(), windowsPrivateKeyCredential);
@@ -210,7 +220,7 @@ public class ComputeEngineCloudWindowsIT {
         assertEquals(logs(), 1, planned.size());
 
         String name = planned.iterator().next().displayName;
-        
+
         // Wait for the node creation to finish
         planned.iterator().next().future.get();
 
@@ -238,7 +248,7 @@ public class ComputeEngineCloudWindowsIT {
         cloud.addConfiguration(snapshotInstanceConfiguration());
 
         FreeStyleProject project = r.createFreeStyleProject();
-        Builder step = new Shell("exit 1");
+        Builder step = new BatchFile("exit 1");
         project.getBuildersList().add(step);
         project.setAssignedLabel(new LabelAtom(SNAPSHOT_LABEL));
 
@@ -307,7 +317,6 @@ public class ComputeEngineCloudWindowsIT {
                 bootDiskProjectId,
                 BOOT_DISK_SIZE_GB_STR,
                 true,
-                WINDOWS_USER,
                 "",
                 windowsPrivateKeyCredentialId,
                 createSnapshot,